Sample working with random integer delays
