+++ src/workloads/docs/CrudFSMTrivial.yml
@@ -18,7 +18,8 @@ Actors:
   Threads: 5
   Database: Test
   Phases:
-  - Repeat: 10  # Run 10 state transition
+  - #Repeat: 10  # Run 10 state transition
+    Duration: 30 seconds
     Collection: Test
     States:
     - Name: First
@@ -32,8 +33,8 @@ Actors:
       - To: Absorbing
         Weight: 1
         #        SleepBefore: &sleep {^TimeSpec: {value: {^RandomDouble: {min: 0, max: 15}}, unit: seconds}}
-        #SleepBefore: &sleep {^TimeSpec: {value: 1, unit: seconds}}
-        SleepBefore: &sleep 1 second
+        SleepBefore: &sleep {^TimeSpec: {value: {^RandomInt: {min: 1, max: 5}}, unit: seconds}}
+        #SleepBefore: &sleep 1 second
       - To: First
         Weight: 1
         SleepBefore: *sleep
@@ -73,8 +74,8 @@ Actors:
       Transitions:
       - To: Absorbing
         Weight: 1
-        SleepBefore: 10 seconds
-        #        SleepBefore: {^TimeSpec: {10000, unit: seconds}}
+        #SleepBefore: 10 seconds
+        SleepBefore: {^TimeSpec: {value: 10, unit: seconds}}
 
     InitialStates:
     - State: First
